+++ src/collections/domain/repositories/ICollectionsRepository.ts
@@ -12,7 +12,7 @@ import { CollectionItemType } from '../../../collections/domain/models/Collectio
 import { CollectionLinks } from '../models/CollectionLinks'
 import { CollectionSummary } from '../models/CollectionSummary'
 import { LinkingObjectType } from '../useCases/GetCollectionsForLinking'
-import { TemplateCreateDTO } from '../dtos/TemplateCreateDTO'
+import { CreateDatasetTemplateDTO } from '../dtos/CreateDatasetTemplateDTO'
 
 export interface ICollectionsRepository {
   getCollection(collectionIdOrAlias: number | string): Promise<Collection>
@@ -71,6 +71,6 @@ export interface ICollectionsRepository {
   ): Promise<CollectionSummary[]>
   createDatasetTemplate(
     collectionIdOrAlias: number | string,
-    template: TemplateCreateDTO
+    template: CreateDatasetTemplateDTO
   ): Promise<void>
 }