+++ src/collections/domain/useCases/CreateDatasetTemplate.ts
@@ -1,7 +1,7 @@
 import { ROOT_COLLECTION_ID } from '../models/Collection'
 import { UseCase } from '../../../core/domain/useCases/UseCase'
 import { ICollectionsRepository } from '../repositories/ICollectionsRepository'
-import { TemplateCreateDTO } from '../dtos/TemplateCreateDTO'
+import { CreateDatasetTemplateDTO } from '../dtos/CreateDatasetTemplateDTO'
 
 export class CreateDatasetTemplate implements UseCase<void> {
   private collectionsRepository: ICollectionsRepository
@@ -13,13 +13,13 @@ export class CreateDatasetTemplate implements UseCase<void> {
   /**
    * Creates a Dataset Template in the specified collection.
    *
-   * @param {TemplateCreateDTO} template - Template definition payload.
+   * @param {CreateDatasetTemplateDTO} template - Template definition payload.
    * @param {number | string} [collectionIdOrAlias = ':root'] - A generic collection identifier, which can be either a string (for queries by CollectionAlias), or a number (for queries by CollectionId)
    * If this parameter is not set, the default value is: ':root'.
    * @returns {Promise<void>}
    */
   async execute(
-    template: TemplateCreateDTO,
+    template: CreateDatasetTemplateDTO,
     collectionIdOrAlias: number | string = ROOT_COLLECTION_ID
   ): Promise<void> {
     return await this.collectionsRepository.createDatasetTemplate(collectionIdOrAlias, template)