+++ src/collections/infra/repositories/CollectionsRepository.ts
@@ -40,7 +40,7 @@ import { ReadError } from '../../../core/domain/repositories/ReadError'
 import { CollectionLinks } from '../../domain/models/CollectionLinks'
 import { CollectionSummary } from '../../domain/models/CollectionSummary'
 import { LinkingObjectType } from '../../domain/useCases/GetCollectionsForLinking'
-import { TemplateCreateDTO } from '../../domain/dtos/TemplateCreateDTO'
+import { CreateDatasetTemplateDTO } from '../../domain/dtos/CreateDatasetTemplateDTO'
 
 export interface NewCollectionRequestPayload {
   alias: string
@@ -532,7 +532,7 @@ export class CollectionsRepository extends ApiRepository implements ICollections
 
   public async createDatasetTemplate(
     collectionIdOrAlias: number | string,
-    template: TemplateCreateDTO
+    template: CreateDatasetTemplateDTO
   ): Promise<void> {
     return this.doPost(
       `/${this.collectionsResourceName}/${collectionIdOrAlias}/templates`,